less -R seems less (hah) problematic than -r
man less:

       -r or --raw-control-chars
              Causes "raw" control characters to be displayed.  The
default is to display control characters using the caret notation; for
example, a control-A (octal 001) is displayed as "^A".  Warning: when
the -r option is used, less cannot keep track of the actual appearance
of the screen (since this depends  on  how  the
              screen responds to each type of control character).  Thus,
various display problems may result, such as long lines being split in
the wrong place.

       -R or --RAW-CONTROL-CHARS
              Like -r, but only ANSI "color" escape sequences are output
in "raw" form.  Unlike -r, the screen appearance is maintained correctly
in most cases.  ANSI "color" escape sequences are sequences of the form:

                   ESC [ ... m

              where  the  "..."  is  zero or more color specification
characters For the purpose of keeping track of screen appearance, ANSI
color escape sequences are assumed to not move the cursor.  You can make
less think that characters other than "m" can end ANSI color escape
sequences by setting the environment variable
              LESSANSIENDCHARS to the list of characters which can end a
color escape sequence.  And you can make less think that characters
other than the standard ones may appear between the ESC and the m by
setting the environment variable LESSANSIMIDCHARS to the list of
characters which can appear.