Replace vim-eunuch with rename.vim again
I got very annoyed by having to think about the difference between :Move and :Rename. It could not AFAIK create missing dirs as part of the rename – or perhaps it did, and I ended up with both the old and the new file. Either way, I did not enjoy it :/